Add assert_set_eq macro to make test clearer
